# src/auth/dto/refresh.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

export class RefreshDto {
  @ApiProperty({
    example: 'b7Zk…token opaco de 64 caracteres…',
    description:
      'El refresh token emitido en login/register o en un refresh anterior',
  })
  @IsString()
  @IsNotEmpty({ message: 'El refresh token es obligatorio' })
  refreshToken: string;
}